Drop 'username' option, use and add verbose flag
The username option may return, but for now it seems that the right way
to configure this utlity is to connect as the current user or use
sshd_config(5) to set User to connect as.

If the '-l' flag is specified '-v' will show the curent options for each
label.