Simplify dry-run and remove list option
Early on '-l' was used in development to verify that options
(interpreter=, execute_with=, ...) were inherited correctly.

In practice, the ability to display these options does not demonstrate
that execution is handled properly since the dry-run ('-n') option takes
a different code path.

If users want the ability to print an execution plan we should come up
with a more comprehensive display log/format.

In the future '-l' option may be re-appropriated to specify a log
directory.