Add archive and restore options
Transfer files using scp(1). The purpose of this is to allow easy backup
of important files from a remote host that should be installed after a
rebuild.

Previously export paths were only allowed in the routes file. Now they
may be added to other labels, where they signify an artifact that may be
archived or restored:

_download: certs.tar
    tar -cpf $SD/certs.tar \
        /etc/ssl/private \
        /etc/ssl/sidecomment.io.*

_upload: certs.tar
    tar -C / -xpf $SD/certs.tar

In this examples the labels names begin with a non-alphanumeric
character so they are not executed by default.

  rset -A -x _download svc1.sidecomment.io

Errors reported by scp(1) are not fatal unless `-e` is specified.

Parallel operation is not supported yet, but this can be be accomplished
by extending custom logging.